requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory_boy==3.3.1
Faker==28.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
rpds-py==0.19.1
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
psycopg2-binary
python-dotenv==1.0.1
django-elasticsearch-dsl~=8.0

graphene~=3.3
graphene-django~=3.2.2